requirements.txt
aiogram==3.0.0b7
aiohttp==3.8.6
python-dotenv==1.0.0
aiofiles==23.1.0